backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
black==22.12.0
Brotli==1.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==4.8.0
Django==4.2.2
django-cors-headers==4.1.0
django-ninja==0.22.1
exceptiongroup==1.2.0
Faker==18.10.1
filetype==1.2.0
frozenlist==1.4.0
har2case==0.3.1
httprunner==2.5.7
idna==3.4
iniconfig==2.0.0
Jinja2==2.11.3
jmespath==0.9.5
jsonpath==0.82
jsonschema==3.2.0
kombu==5.3.1
loguru==0.4.1
MarkupSafe==1.1.1
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pluggy==1.3.0
prompt-toolkit==3.0.39
py==1.11.0
pycrypto==2.6.1
pydantic==1.8.2
PyMySQL==1.0.3
pyrsistent==0.19.3
pystache==0.6.0
pytest==7.4.3
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
PyYAML==5.1.2
redis==4.6.0
requests==2.31.0
requests-toolbelt==0.9.1
sentry-sdk==0.13.5
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.18
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.3
